+++ changelog.md
@@ -1,4 +1,7 @@
 # Changelog
+## 3.6.0
+- add [copy results from Chatlog](https://github.com/AngryBeaver/beavers-crafting/issues/143)
+  - configurable via Settings
 ## 3.5.5
 - fix compatibility to pf1
 - removed scrollbars when not needed